test view helpers
