+++ packages/view/test/fixtures/resources/views/test-view-helper-json-pretty.hbs
@@ -1,0 +1,1 @@
+{{json user pretty=true}}
