+++ packages/view/test/fixtures/resources/views/test-view-helper-json.hbs
@@ -1,0 +1,1 @@
+<p>{{json user}}</p>
