+++ packages/view/test/fixtures/resources/views/test-view-helper-stack-append-prepend.hbs
@@ -1,0 +1,11 @@
+{{#append "test-stack"}}
+  append
+{{/append}}
+
+{{#prepend "test-stack"}}
+  prepend
+{{/prepend}}
+
+{{#stack "test-stack"}}
+  content
+{{/stack}}
